chore:card scripts 파일 분할
