+++ src/client/scripts/card-handlers/index.ts
@@ -1,0 +1,3 @@
+export * from './create-a-card-handlers'
+export * from './modify-a-card-handlers'
+export * from './remove-a-card-handlers'
